Re-use validator for value
