+++ vizro-core/src/vizro/models/_components/form/_form_utils.py
@@ -1,6 +1,8 @@
 """Helper functions for models inside form folder."""
+from typing import Union
+
 from vizro._constants import ALL_OPTION
-from vizro.models.types import OptionsType
+from vizro.models.types import MultiValueType, OptionsType, SingleValueType
 
 
 def get_options_and_default(options: OptionsType, multi: bool = False):
@@ -19,9 +21,17 @@ def get_options_and_default(options: OptionsType, multi: bool = False):
     return options, default_value
 
 
+def is_value_contained(value: Union[SingleValueType, MultiValueType], options: OptionsType):
+    """Checks if value is contained in a list."""
+    if isinstance(value, list):
+        return all(item in options for item in value)
+    else:
+        return value in options
+
+
 # Validators for re-use
 def validate_options_dict(cls, values):
-    """Reusable validator for "options" argument of categorical selectors."""
+    """Reusable validator for the "options" argument of categorical selectors."""
     if "options" not in values or not isinstance(values["options"], list):
         return values
 
@@ -31,3 +41,18 @@ def validate_options_dict(cls, values):
                 "Invalid argument `options` passed into Selector. Expected a dict with keys `label` and `value`."
             )
     return values
+
+
+def validate_value(cls, value, values):
+    """Reusable validator for the "value" argument of categorical selectors."""
+    if "options" not in values or not values["options"]:
+        return value
+
+    possible_values = (
+        [entry["value"] for entry in values["options"]] if isinstance(values["options"][0], dict) else values["options"]
+    )
+
+    if value and not is_value_contained(value, possible_values):
+        raise ValueError("Please provide a valid value from `options`.")
+
+    return value
